requirements.txt
about-time==3.1.1
appdirs==1.4.4
astroid==2.4.2
bleach==4.1.0
build==0.7.0
certifi==2020.6.20
cffi==1.15.0
charset-normalizer==2.0.9
colorama==0.4.4
cryptography==36.0.1
distlib==0.3.1
docutils==0.18.1
filelock==3.0.12
grapheme==0.6.0
idna==3.3
importlib-metadata>=4.8.3
isort==5.7.0
jeepney==0.7.1
keyring==23.4.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
packaging==21.3
pep517==0.12.0
pkginfo==1.8.2
polib==1.1.1
progressbar2==3.55.0
pycparser==2.21
Pygments==2.10.0
pylint==2.6.0
pyparsing==3.0.6
python-utils==2.5.6
readme-renderer==32.0
requests==2.26.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
SecretStorage==3.3.1
six==1.16.0
toml==0.10.2
tomli>=1.2.3
tqdm==4.62.3
twine==3.7.1
urllib3==1.26.7
virtualenv==20.4.2
webencodings==0.5.1
wrapt==1.12.1
zipp==3.6.0
